# docs/swagger/admin-api.yml
openapi: 3.1.0
info:
  title: Admin API
  description: Library of Coffee - Admin API
  contact:
    name: Library of Coffee
  version: 1.0.0
servers:
- url: /
  description: Current Server
security:
- Bearer: []
...
paths:
...
  /api/admin/coffee-beans:
    get:
      tags:
      - CoffeeBean
      summary: コーヒー豆一覧取得
      description: コーヒー豆の一覧をページネーション付きで取得します。画像・テイスト評価は含まれません。
      operationId: listCoffeeBeans
      parameters:
      - name: page
        in: query
        description: ページ番号（0始まり）
        required: false
        schema:
          type: integer
          format: int32
          default: 0
        example: 0
      - name: size
        in: query
        description: 1ページあたりの件数
        required: false
        schema:
          type: integer
          format: int32
          default: 20
        example: 20
      responses:
        "200":
          description: 取得成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CoffeeBeanListResponse"
              examples:
                success:
                  summary: 取得成功例
                  description: success
                  value:
                    items:
                    - id: 00000000-0000-4000-8000-000000000001
                      shopId: 00000000-0000-4000-8000-000000000002
                      shopName: コーヒーショップ青山
                      shopifyBeanId: test-bean-001
                      name: エチオピア イルガチェフェ
                      description: フルーティーな香りが特徴的なコーヒー豆です。
                      origin: エチオピア
                      farm: イルガチェフェ農園
                      roastLevel: MEDIUM
                      processingMethod: WASHED
                      isSpecialty: true
                      publishStatus: PUBLISHED
                    totalCount: 1
                    page: 0
                    size: 20
    post:
      tags:
      - CoffeeBean
      summary: コーヒー豆登録
      description: 新しいコーヒー豆を登録します。multipart/form-data形式で送信してください。CoffeeBeanId、CoffeeBeanImageId、CoffeeBeanTasteIdはサーバー側でUUIDv4を自動生成します。
      operationId: createCoffeeBean
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  $ref: "#/components/schemas/CreateCoffeeBeanRequest"
                images:
                  type: array
                  items:
                    type: string
                    format: binary
              required:
              - data
      responses:
        "201":
          description: 登録成功
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CoffeeBeanResponse"
              examples:
                success:
                  summary: 登録成功例
                  description: success
                  value:
                    id: 00000000-0000-4000-8000-000000000001
        "400":
          description: バリデーションエラー
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                bad_request:
                  summary: バリデーションエラーの場合
                  description: bad_request
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 400
                    error: Bad Request
                    path: /api/admin/coffee-beans
        "409":
          description: Shopify商品IDが重複しています
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                conflict:
                  summary: Shopify商品IDが重複している場合
                  description: conflict
                  value:
                    timestamp: 2026-02-23T12:00:00.000+00:00
                    status: 409
                    error: Conflict
                    path: /api/admin/coffee-beans
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: エラーレスポンス
      properties:
        timestamp:
          type: string
          format: date-time
          description: エラー発生日時
          example: 2026-02-23T12:00:00.000+00:00
        status:
          type: integer
          format: int32
          description: HTTPステータスコード
          example: 404
        error:
          type: string
          description: エラー概要
          example: Not Found
        message:
          type: string
          description: エラーメッセージ
          example: tastes must not be empty
        path:
          type: string
          description: リクエストパス
          example: /api/admin/resources/00000000-0000-4000-8000-000000000099
      required:
      - error
      - path
      - status
      - timestamp
...
    TasteRequest:
      type: object
      description: テイスト評価リクエスト
      properties:
        tasteId:
          type: string
          description: テイストID
          example: 00000000-0000-4000-8000-000000000041
        evaluationValue:
          type: integer
          format: int32
          description: 評価値（0-5）
          example: 3
      required:
      - evaluationValue
      - tasteId
...
    CoffeeBeanResponse:
      type: object
      description: コーヒー豆登録レスポンス
      properties:
        id:
          type: string
          description: コーヒー豆ID
          example: 00000000-0000-4000-8000-000000000001
      required:
      - id
...
    CreateCoffeeBeanRequest:
      type: object
      description: コーヒー豆登録リクエスト
      properties:
        shopId:
          type: string
          description: 店舗ID
          example: 00000000-0000-4000-8000-000000000031
        shopifyBeanId:
          type: string
          description: Shopifyの商品ID
          example: gid://shopify/Product/999999
        name:
          type: string
          description: 豆の名前
          example: テストコーヒー豆
        description:
          type: string
          description: 説明
          example: テスト説明文
        origin:
          type: string
          description: 産地
          example: エチオピア
        farm:
          type: string
          description: 農園名
          example: テスト農園
        roastLevel:
          type: string
          description: 焙煎度
          example: MEDIUM
        processingMethod:
          type: string
          description: 精製方法
          enum:
          - FULLY_WASHED
          - WASHED
          - ANAEROBIC_WASHED
          - THERMAL_SHOCK_NATURAL
          - NATURAL
          - ANAEROBIC_NATURAL
          - DRY_ON_TREE_NATURAL
          - LACTIC_NATURAL
          - WET_HULLING
          - HONEY
          - MOUNTAIN_WATER
          - LADO_A_LADO_PROCESS
          - LADO_A_LADO_PROCESS_FULLY_WASHED
          example: WASHED
        publishStatus:
          type: string
          description: "公開状態（DRAFT: 下書き / PUBLISHED: 公開）"
          example: DRAFT
        tastes:
          type: array
          description: テイスト評価一覧
          items:
            $ref: "#/components/schemas/TasteRequest"
        imageTypes:
          type: array
          description: アップロード画像のタイプ一覧（imagesと同じ順序）
          example:
          - MAIN
          items:
            type: string
        isSpecialty:
          type: boolean
          description: スペシャルティコーヒーかどうか
          example: true
      required:
      - description
      - isSpecialty
      - name
      - origin
      - processingMethod
      - publishStatus
      - roastLevel
      - shopId
      - shopifyBeanId
      - tastes
...
    CoffeeBeanListResponse:
      type: object
      description: コーヒー豆一覧レスポンス
      properties:
        items:
          type: array
          description: アイテム一覧
          items:
            $ref: "#/components/schemas/CoffeeBeanSummaryResponse"
        totalCount:
          type: integer
          format: int64
          description: 全件数
          example: 42
        page:
          type: integer
          format: int32
          description: 現在のページ番号（0始まり）
          example: 0
        size:
          type: integer
          format: int32
          description: 1ページあたりの件数
          example: 20
      required:
      - items
      - page
      - size
      - totalCount
    CoffeeBeanSummaryResponse:
      type: object
      description: コーヒー豆一覧アイテム
      properties:
        id:
          type: string
          description: コーヒー豆ID
          example: 00000000-0000-4000-8000-000000000001
        shopId:
          type: string
          description: ショップID
          example: 00000000-0000-4000-8000-000000000002
        shopName:
          type: string
          description: 店舗名
          example: コーヒーショップ青山
        shopifyBeanId:
          type: string
          description: Shopify商品ID
          example: test-bean-001
        name:
          type: string
          description: 豆の名前
          example: エチオピア イルガチェフェ
        description:
          type: string
          description: 説明
          example: フルーティーな香りが特徴的なコーヒー豆です。
        origin:
          type: string
          description: 産地
          example: エチオピア
        farm:
          type: string
          description: 農園名
          example: イルガチェフェ農園
        roastLevel:
          type: string
          description: 焙煎度
          example: MEDIUM
        processingMethod:
          type: string
          description: 精製方法
          enum:
          - FULLY_WASHED
          - WASHED
          - ANAEROBIC_WASHED
          - THERMAL_SHOCK_NATURAL
          - NATURAL
          - ANAEROBIC_NATURAL
          - DRY_ON_TREE_NATURAL
          - LACTIC_NATURAL
          - WET_HULLING
          - HONEY
          - MOUNTAIN_WATER
          - LADO_A_LADO_PROCESS
          - LADO_A_LADO_PROCESS_FULLY_WASHED
          example: WASHED
        processingMethodName:
          type: string
          description: 精製方法の表示名
          example: ウォッシュド
        publishStatus:
          type: string
          description: "公開状態（DRAFT: 下書き / PUBLISHED: 公開）"
          example: PUBLISHED
        isSpecialty:
          type: boolean
          description: スペシャルティコーヒーかどうか
          example: true
      required:
      - description
      - id
      - isSpecialty
      - name
      - origin
      - processingMethod
      - processingMethodName
      - publishStatus
      - roastLevel
      - shopId
      - shopName
      - shopifyBeanId
...
  securitySchemes:
    Bearer:
      type: http
      description: JWTアクセストークン
      scheme: bearer
      bearerFormat: JWT